Fix: change long to int32_t for purchases and last stinky chocolate time in PlayerInfo
* Ensure the behavior with big endian

+++ src/Lawn/System/PlayerInfo.h
@@ -4,6 +4,7 @@
 #define MAX_POTTED_PLANTS 200
 #define PURCHASE_COUNT_OFFSET 1000
 
+#include <cstdint>
 #include <ctime>
 #include <vector>
 #include "../../ConstEnums.h"
@@ -52,13 +53,13 @@ public:
     int                 mCoins;                             //+0x28
     int                 mFinishedAdventure;                 //+0x2C
     int                 mChallengeRecords[100];             //+0x30
-    long                mPurchases[80];                     //+0x1C0
+    int32_t             mPurchases[80];                     //+0x1C0
     int                 mPlayTimeActivePlayer;              //+0x300
     int                 mPlayTimeInactivePlayer;            //+0x304
     int                 mHasUsedCheatKeys;                  //+0x308
     int                 mHasWokenStinky;                    //+0x30C
     int                 mDidntPurchasePacketUpgrade;        //+0x310
-    long                mLastStinkyChocolateTime;           //+0x314
+    int32_t             mLastStinkyChocolateTime;           //+0x314
     int                 mStinkyPosX;                        //+0x318
     int                 mStinkyPosY;                        //+0x31C
     int                 mHasUnlockedMinigames;              //+0x320
